% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manaakiwhenua-package.R
\docType{package}
\name{manaakiwhenua}
\alias{manaakiwhenua}
\alias{manaakiwhenua-package}
\title{Themes and Graphics Resources for Manaaki Whenua}
\description{
A compilation of colour palettes and 'ggplot2' themes and scales 
for the new Manaaki Whenua branding scheme.
}
\author{
Pierre Roudier (roudierp@landcareresearch.co.nz)
}
