% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{mw_colours}
\alias{mw_colours}
\title{Manaaki Whenua corporate colours}
\format{An object of class \code{character} of length 9.}
\usage{
mw_colours
}
\description{
A vector of the Manaaki Whenua corporate colours. The vector is named so to be able to refer to colours by name.
}
\author{
Pierre Roudier, from the "Manaaki Whenua colour theme" page of the brand toolkit.
}
\keyword{datasets}
