% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{mw_cols}
\alias{mw_cols}
\title{Extract Manaaki Whenua colours as HEX}
\usage{
mw_cols(...)
}
\arguments{
\item{...}{Either the character names of of the colours available in `mw_colors`, or integer(s) giving their position in `mw_colours`.}
}
\description{
Function to extract the Manaaki Whenua colours as HEX codes
}
\author{
Pierre Roudier
}
