% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{mw_pal}
\alias{mw_pal}
\title{Return function to interpolate a Manaaki Whenua colour palette}
\usage{
mw_pal(palette = "dark", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in mw_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
