% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{mw_palettes}
\alias{mw_palettes}
\title{Manaaki Whenua colour palettes}
\format{An object of class \code{list} of length 4.}
\usage{
mw_palettes
}
\description{
List of colours for the Manaaki Whenua palettes
}
\author{
Pierre Roudier
}
\keyword{datasets}
