% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_mw}
\alias{scale_colour_mw}
\alias{scale_color_mw}
\title{Colour scale constructor from the Manaaki Whenua colour palettes}
\usage{
scale_colour_mw(palette = "dark", discrete = TRUE, reverse = FALSE, ...)

scale_color_mw(palette = "dark", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of the palette (see `mw_palettes`)}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
